requirements.txt
discord.py
dotenv
openpipe
aiohttp
pytz